requirements.txt
# Core scientific stack
numpy>=1.21.0
scipy>=1.7.0

# Configuration
pyyaml>=6.0

# Plotting (for notebooks)
matplotlib>=3.5.0
seaborn>=0.11.0

# Metrics
scikit-image>=0.19.0    # for SSIM
networkx>=2.6           # for modularity

# Statistics
scipy>=1.7.0

# VAE (only needed for steps 07-08)
tensorflow>=2.8.0

# Optional but useful
pandas>=1.3.0           # for tables in notebooks
